renamed dialog file
