+++ KP_ui.hpp
@@ -4,7 +4,7 @@
     File: KP_ui.hpp
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2017-10-16
-    Last Update: 2019-08-07
+    Last Update: 2020-02-05
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
@@ -12,5 +12,5 @@
 */
 
 #include "KPGUI\KPGUI_defines.hpp"
-#include "KPCF\ui\KPCF_dialog.hpp"
+#include "KPCF\ui\KP_cratefiller_dialog.hpp"
 #include "KPCF\ui\defines.hpp"
